# docs/openapi/product-api-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "EventListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "title": "EventListResponse",
        "type": "object"
      },
...
      "ResidentAnalysisListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResidentAnalysisResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "title": "ResidentAnalysisListResponse",
        "type": "object"
      },
...
      "ResidentListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResidentSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "title": "ResidentListResponse",
        "type": "object"
      },
...
      "ResidentSummary": {
        "additionalProperties": false,
        "properties": {
          "assignment_status": {
            "const": "active",
            "title": "Assignment Status",
            "type": "string"
          },
          "display_label": {
            "title": "Display Label",
            "type": "string"
          },
          "resident_id": {
            "title": "Resident Id",
            "type": "string"
          },
          "room_id": {
            "title": "Room Id",
            "type": "string"
          },
          "room_label": {
            "title": "Room Label",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "resident_id",
          "display_label",
          "room_id",
          "room_label",
          "assignment_status"
        ],
        "title": "ResidentSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Contactless Monitoring Product API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/residents": {
      "get": {
        "operationId": "list_residents_v1_residents_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Residents",
        "tags": [
          "residents"
        ]
      }
    },
    "/v1/residents/{resident_id}": {
      "get": {
        "operationId": "get_resident_v1_residents__resident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "resident_id",
            "required": true,
            "schema": {
              "title": "Resident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Resident",
        "tags": [
          "residents"
        ]
      }
    },
    "/v1/residents/{resident_id}/analyses": {
      "get": {
        "operationId": "list_resident_analyses_v1_residents__resident_id__analyses_get",
        "parameters": [
          {
            "in": "path",
            "name": "resident_id",
            "required": true,
            "schema": {
              "title": "Resident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentAnalysisListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Resident Analyses",
        "tags": [
          "residents"
        ]
      }
    },
...
    "/v1/residents/{resident_id}/events": {
      "get": {
        "operationId": "list_resident_events_v1_residents__resident_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "resident_id",
            "required": true,
            "schema": {
              "title": "Resident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Resident Events",
        "tags": [
          "residents"
        ]
      }
    },
    "/v1/residents/{resident_id}/memory": {
      "get": {
        "operationId": "get_resident_memory_v1_residents__resident_id__memory_get",
        "parameters": [
          {
            "in": "path",
            "name": "resident_id",
            "required": true,
            "schema": {
              "title": "Resident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentMemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Resident Memory",
        "tags": [
          "residents"
        ]
      }
    },
    "/v1/residents/{resident_id}/memory/entries": {
      "post": {
        "operationId": "add_resident_memory_entry_v1_residents__resident_id__memory_entries_post",
        "parameters": [
          {
            "in": "path",
            "name": "resident_id",
            "required": true,
            "schema": {
              "title": "Resident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemoryEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentMemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Add Resident Memory Entry",
        "tags": [
          "residents"
        ]
      }
    },
    "/v1/residents/{resident_id}/memory/entries/{entry_id}/correct": {
      "post": {
        "operationId": "correct_resident_memory_entry_v1_residents__resident_id__memory_entries__entry_id__correct_post",
        "parameters": [
          {
            "in": "path",
            "name": "resident_id",
            "required": true,
            "schema": {
              "title": "Resident Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectMemoryEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentMemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Correct Resident Memory Entry",
        "tags": [
          "residents"
        ]
      }
    },
    "/v1/residents/{resident_id}/memory/entries/{entry_id}/retire": {
      "post": {
        "operationId": "retire_resident_memory_entry_v1_residents__resident_id__memory_entries__entry_id__retire_post",
        "parameters": [
          {
            "in": "path",
            "name": "resident_id",
            "required": true,
            "schema": {
              "title": "Resident Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetireMemoryEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentMemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Retire Resident Memory Entry",
        "tags": [
          "residents"
        ]
      }
    },
...
  }
}